Old/translatePP/requirements.txt
python-pptx>=0.6.21
python-docx>=0.8.11
PyMuPDF>=1.23.8
openai>=1.12.0
python-dotenv>=1.0.0
tqdm>=4.66.1 